Add admin profile page with updae infromation and delete profile possibilitys, add full role based accesse functionality in front and back-end

+++ app/Http/Controllers/EventController.php
@@ -14,8 +14,12 @@ class EventController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Event::class) && ($request->user()->cannot('create', Event::class) || $request->user()->cannot('delete', Event::class) || $request->user()->cannot('update', Event::class))) {
+            return Inertia::render('Error/Error_403');
+        }
+
         $events = Event::with('assets')->get();
         return Inertia::render('Admin/Events/Events', ['events' => $events]);
     }
@@ -23,8 +27,11 @@ class EventController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        if ($request->user()->cannot('create', Event::class)) {
+            return Inertia::render('Error/Error_403');
+        }
         return Inertia::render('Admin/Events/CreateEvent');
     }
 
@@ -33,6 +40,9 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->user()->cannot('create', Event::class)) {
+            return Inertia::render('Error/Error_403');
+        }
         request()->validate(
             [
                 'event_name' => 'required|string',
@@ -76,8 +86,11 @@ class EventController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(string $id, Request $request)
     {
+        if ($request->user()->cannot('update', Event::class)) {
+            return Inertia::render('Error/Error_403');
+        }
         $event = Event::with('assets')->where('event_id', $id)->first();
         return Inertia::render('Admin/Events/EditEvent', ['event' => $event]);
     }
@@ -87,6 +100,9 @@ class EventController extends Controller
      */
     public function update(Request $request)
     {
+        if ($request->user()->cannot('update', Event::class)) {
+            return Inertia::render('Error/Error_403');
+        }
         request()->validate(
             [
                 'event_name' => 'required|string',
@@ -132,8 +148,11 @@ class EventController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, Request $request)
     {
+        if ($request->user()->cannot('delete', Event::class)) {
+            return Inertia::render('Error/Error_403');
+        }
         Event::where('event_id', $id)->delete();
         return redirect()->back()->with('message', ['status' => 'success', 'message'
         => 'Evènement supprimé avec succès']);
